Set up cooldown
Set up cooldown after using a shout. Don't allow the user to switch shouts if they are charging.

Fix a bug that prevented the user from using whirlwind sprint after using unrelenting force and not hitting an enemy.

Mess around with settings in the spell effects to test what the values do